+++ Classes/Emogrifier.php
@@ -155,17 +155,6 @@ class Emogrifier
     private $shouldKeepInvisibleNodes = true;
 
     /**
-     * Determines whether media query style blocks should be added to the body versus the head of the HTML.
-     * 
-     * If set to true, the <style> blocks will be added to the <body>
-     * 
-     * If set to false, the <style> blocks will be added to the <head>
-     *
-     * @var bool
-     */
-    private $shouldAddStyleToBody = false;
-
-    /**
      * @var string[]
      */
     private $xPathRules = [
@@ -667,16 +656,6 @@ class Emogrifier
     }
 
     /**
-     * Adds media query styles to <body> instead of <head>
-     *
-     * @return void
-     */
-    public function addStyleToBody()
-    {
-        $this->shouldAddStyleToBody = true;
-    }
-
-    /**
      * Clears all caches.
      *
      * @return void
@@ -1058,19 +1037,9 @@ class Emogrifier
         $styleAttribute->value = 'text/css';
         $styleElement->appendChild($styleAttribute);
 
-        if ($this->shouldAddStyleToBody) {
-            $body = $this->getBodyElement($document);
+        $body = $this->getBodyElement($document);
+        if ($body) {
             $body->appendChild($styleElement);
-            if ($body->hasChildNodes()) {
-                $body->insertBefore($styleElement,$body->firstChild);
-            }
-            else {
-                $body->appendChild($styleElement);
-            }
-        }
-        else {
-            $head = $this->getOrCreateHeadElement($document);
-            $head->appendChild($styleElement);
         }
     }
 
@@ -1103,9 +1072,7 @@ class Emogrifier
      */
     private function getBodyElement(\DOMDocument $document)
     {
-        $body = $document->getElementsByTagName('body')->item(0);
-
-        return $body;
+        return $document->getElementsByTagName('body')->item(0);
     }
 
     /**